requirements.txt
importlib-metadata # from an error message
ncbi-datasets-cli
biopython
khmer
numpy
pandas
pytest
screed
seaborn
ipython
